fix: move actions to MoreOptions
