+++ services/orchest-webserver/client/src/jobs-view/edit-job/JobPrimaryButtonIcon.tsx
@@ -1,6 +1,6 @@
 import CancelOutlinedIcon from "@mui/icons-material/CancelOutlined";
 import ContentCopyOutlinedIcon from "@mui/icons-material/ContentCopyOutlined";
-import PlayCircleOutlineOutlinedIcon from "@mui/icons-material/PlayCircleOutlineOutlined";
+import PlayArrowOutlinedIcon from "@mui/icons-material/PlayArrowOutlined";
 import ReplayOutlinedIcon from "@mui/icons-material/ReplayOutlined";
 import ScheduleOutlinedIcon from "@mui/icons-material/ScheduleOutlined";
 import StopCircleOutlinedIcon from "@mui/icons-material/StopCircleOutlined";
@@ -8,7 +8,7 @@ import Box from "@mui/material/Box";
 import React from "react";
 
 const iconMapping = {
-  run: <PlayCircleOutlineOutlinedIcon fontSize="small" />,
+  run: <PlayArrowOutlinedIcon fontSize="small" />,
   schedule: <ScheduleOutlinedIcon fontSize="small" />,
   pause: <StopCircleOutlinedIcon fontSize="small" />,
   resume: <ReplayOutlinedIcon fontSize="small" />,
