+++ services/orchest-webserver/client/src/jobs-view/edit-job/hooks/useJobPrimaryButtonActions.tsx
@@ -5,58 +5,30 @@ import { JobPrimaryButtonIconType } from "../JobPrimaryButtonIcon";
 import { useJobActions } from "./useJobActions";
 
 export const useJobPrimaryButtonActions = () => {
-  const {
-    scheduleJob,
-    resumeJob,
-    pauseJob,
-    cancelJob,
-    duplicateJob,
-  } = useJobActions();
+  const { scheduleJob, cancelJob, duplicateJob } = useJobActions();
 
   const status = useEditJob((state) => state.jobChanges?.status);
   const schedule = useEditJob((state) => state.jobChanges?.schedule);
-  const nextScheduledTime = useEditJob(
-    (state) => state.jobChanges?.next_scheduled_time
-  );
 
-  const hasStarted = status === "STARTED" || status === "PENDING";
+  const hasStarted =
+    status === "STARTED" || status === "PENDING" || status === "PAUSED";
 
   const [buttonLabel, mainAction, iconType] = React.useMemo<
     [string, () => void, JobPrimaryButtonIconType]
   >(() => {
-    const isScheduledJob = hasValue(schedule) || hasValue(nextScheduledTime);
-
-    const hasPaused = status === "PAUSED";
-
     if (status === "DRAFT") {
-      const shouldRunNow = !hasValue(schedule) && !hasValue(nextScheduledTime);
+      const isCronJob = hasValue(schedule);
 
-      if (shouldRunNow) return ["Run job", scheduleJob, "run"];
-      return ["Schedule job", scheduleJob, "schedule"];
+      if (isCronJob) return ["Schedule job", scheduleJob, "schedule"];
+      return ["Run job", scheduleJob, "run"];
     }
 
-    if (hasPaused) {
-      return ["Resume job", resumeJob, "resume"];
-    }
-    if (isScheduledJob && hasStarted) {
-      return ["Pause job", pauseJob, "pause"];
-    }
-    if (!isScheduledJob && hasStarted) {
+    if (hasStarted) {
       return ["Cancel job", cancelJob, "cancel"];
     }
 
     return ["Copy job configuration", duplicateJob, "duplicate"];
-  }, [
-    nextScheduledTime,
-    schedule,
-    status,
-    hasStarted,
-    scheduleJob,
-    resumeJob,
-    pauseJob,
-    cancelJob,
-    duplicateJob,
-  ]);
+  }, [schedule, status, hasStarted, scheduleJob, cancelJob, duplicateJob]);
 
   return [buttonLabel, mainAction, iconType] as const;
 };
